+++ src/Components/CreateImageWizard/steps/Oscap/components/ProfileSelector.tsx
@@ -30,7 +30,6 @@ import {
   DistributionProfileItem,
   Filesystem,
   OpenScapProfile,
-  Services,
 } from '../../../../../store/imageBuilderApi';
 import {
   changeCompliance,
@@ -39,9 +38,6 @@ import {
   addPartition,
   changeFileSystemConfigurationType,
   clearPartitions,
-  changeEnabledServices,
-  changeMaskedServices,
-  changeDisabledServices,
   selectComplianceType,
   clearKernelAppend,
   addKernelArg,
@@ -107,7 +103,8 @@ const ProfileSelector = () => {
   const [selectOptions, setSelectOptions] = useState<string[]>([]);
   const complianceType = useAppSelector(selectComplianceType);
   const prefetchProfile = useBackendPrefetch('getOscapCustomizations');
-  const { clearCompliancePackages, handlePackages } = useSelectorHandlers();
+  const { clearCompliancePackages, handlePackages, handleServices } =
+    useSelectorHandlers();
 
   const {
     data: profiles,
@@ -205,12 +202,6 @@ const ProfileSelector = () => {
         dispatch(addPartition(partition));
       }
     }
-  };
-
-  const handleServices = (services: Services | undefined) => {
-    dispatch(changeEnabledServices(services?.enabled || []));
-    dispatch(changeMaskedServices(services?.masked || []));
-    dispatch(changeDisabledServices(services?.disabled || []));
   };
 
   const handleKernelAppend = (kernelAppend: string | undefined) => {
